+++ app/src/main/res/layout/activity_signup.xml
@@ -1,0 +1,185 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    android:background="#0a0c16"
+    tools:context=".activities.signupActivity">
+
+    <ImageView
+        android:layout_width="50dp"
+        android:layout_height="50dp"
+        android:layout_margin="15dp"
+        android:src="@drawable/back"/>
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Sign Up"
+        android:textColor="@color/white"
+        android:fontFamily="@font/poppins"
+        android:layout_gravity="center"
+        android:layout_marginTop="-58dp"
+        android:textSize="30sp"/>
+
+    <LinearLayout
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:orientation="horizontal">
+
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="Create Account"
+            android:textStyle="bold"
+            android:textColor="@color/white"
+            android:fontFamily="@font/poppins"
+            android:layout_marginStart="20dp"
+            android:layout_marginTop="60dp"
+            android:textSize="30sp"/>
+        <ImageView
+            android:layout_width="100dp"
+            android:layout_height="100dp"
+            android:layout_marginStart="50dp"
+            android:src="@drawable/color_balls"/>
+
+    </LinearLayout>
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Please Inter your information and\ncreate your account"
+        android:textStyle="bold"
+        android:textColor="#8B8A8A"
+        android:fontFamily="@font/poppins"
+        android:layout_marginStart="20dp"
+        android:layout_marginTop="5dp"
+        android:textSize="15sp"/>
+
+
+    <com.google.android.material.textfield.TextInputLayout
+        android:layout_width="360dp"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="15dp"
+        android:layout_gravity="center"
+        android:textColorHint="@color/white"
+        app:boxCornerRadiusTopStart="15dp"
+        app:boxCornerRadiusBottomEnd="15dp"
+        app:boxCornerRadiusBottomStart="15dp"
+        app:boxCornerRadiusTopEnd="15dp"
+        android:textColor="@color/white"
+        app:boxStrokeColor="#1540DD"
+        app:hintTextColor="#1540DD"
+        android:background="#0a0c16">
+        <com.google.android.material.textfield.TextInputEditText
+            android:id="@+id/email"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:hint="Enter your name"
+            android:fontFamily="@font/poppins"
+            android:textColorHighlight="@color/white" />
+    </com.google.android.material.textfield.TextInputLayout>
+
+    <com.google.android.material.textfield.TextInputLayout
+        android:layout_width="360dp"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center"
+        android:layout_marginTop="15dp"
+        app:boxCornerRadiusTopStart="15dp"
+        app:boxCornerRadiusBottomEnd="15dp"
+        app:boxCornerRadiusBottomStart="15dp"
+        app:boxStrokeColor="#1540DD"
+        app:boxCornerRadiusTopEnd="15dp"
+        android:textColorHint="@color/white"
+        android:textColor="@color/white"
+        app:hintTextColor="#1540DD"
+        android:background="#0a0c16">
+        <com.google.android.material.textfield.TextInputEditText
+            android:id="@+id/password"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:hint="Enter your mail"
+            android:fontFamily="@font/poppins"
+            android:textColorHighlight="@color/white" />
+    </com.google.android.material.textfield.TextInputLayout>
+
+    <com.google.android.material.textfield.TextInputLayout
+        android:layout_width="360dp"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center"
+        android:layout_marginTop="15dp"
+        app:boxCornerRadiusTopStart="15dp"
+        app:boxCornerRadiusBottomEnd="15dp"
+        app:boxCornerRadiusBottomStart="15dp"
+        app:boxStrokeColor="#1540DD"
+        app:boxCornerRadiusTopEnd="15dp"
+        android:textColorHint="@color/white"
+        android:textColor="@color/white"
+        app:hintTextColor="#1540DD"
+        android:background="#0a0c16">
+        <com.google.android.material.textfield.TextInputEditText
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:hint="Enter your password"
+            android:fontFamily="@font/poppins"
+            android:textColorHighlight="@color/white" />
+    </com.google.android.material.textfield.TextInputLayout>
+
+    <Button
+        android:id="@+id/signupBtn"
+        android:layout_width="350dp"
+        android:layout_height="60dp"
+        android:text="Sign Up"
+        android:textSize="18sp"
+        android:fontFamily="@font/poppins"
+        android:textColor="@color/white"
+        android:backgroundTint="#3580ff"
+        android:layout_gravity="center"
+        android:layout_marginTop="15dp"/>
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Signup With"
+        android:textStyle="bold"
+        android:textColor="#8B8A8A"
+        android:fontFamily="@font/poppins"
+        android:layout_gravity="center"
+        android:layout_marginTop="12dp"
+        android:textSize="15sp"/>
+
+
+    <LinearLayout
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="8dp"
+        android:layout_gravity="center"
+        android:orientation="horizontal">
+
+        <ImageView
+            android:layout_width="70dp"
+            android:layout_height="70dp"
+            android:src="@drawable/apple"/>
+
+        <ImageView
+            android:layout_width="70dp"
+            android:layout_height="70dp"
+            android:layout_marginStart="15dp"
+            android:src="@drawable/google"/>
+
+
+    </LinearLayout>
+
+    <TextView
+        android:id="@+id/signin"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/signupString"
+        android:textStyle="bold"
+        android:textColor="#CBC7C7"
+        android:fontFamily="@font/poppins"
+        android:layout_gravity="center"
+        android:layout_marginTop="20dp"
+        android:textSize="15sp"/>
+
+</LinearLayout>
