+++ app/src/main/res/values/strings.xml
@@ -60,4 +60,6 @@
     <string name="intro2String">Work more <font color="#3580ff"><b>Structured</b></font> and Organized 👌</string>
     <string name="intro3String">Manage your <font color="#3580ff"><b>Tasks</b></font> quickly for Results✌</string>
     <string name="searchHint"><font face="poppins_medium" size="14"><b>Search</b></font></string>
+    <string name="signinString">Not Register Yet? <font color="#3580ff"><b>Sign Up</b></font></string>
+    <string name="signupString">Have an Account? <font color="#3580ff"><b>Sign In</b></font></string>
 </resources>
